Move other V2 tests to new test namespace
