+++ osu.Game.Tests/Visual/SongSelectV2/TestSceneBeatmapInfoWedgeV2.cs
@@ -18,7 +18,7 @@ using osu.Game.Rulesets.Objects.Types;
 using osu.Game.Screens.Select;
 using osuTK;
 
-namespace osu.Game.Tests.Visual.SongSelect
+namespace osu.Game.Tests.Visual.SongSelectV2
 {
     [TestFixture]
     public partial class TestSceneBeatmapInfoWedgeV2 : OsuTestScene
