+++ osu.Game.Tests/Visual/SongSelectV2/TestSceneSongSelectV2Navigation.cs
@@ -7,7 +7,7 @@ using osu.Framework.Testing;
 using osu.Game.Screens.Menu;
 using osuTK.Input;
 
-namespace osu.Game.Tests.Visual.SongSelect
+namespace osu.Game.Tests.Visual.SongSelectV2
 {
     public partial class TestSceneSongSelectV2Navigation : OsuGameTestScene
     {
